+++ code/cgame/cg_info.c
@@ -293,7 +293,7 @@ void CG_DrawInformation( void ) {
 	}
 
 	/* Fraglimits are restricted to non-team-based games and Team Deathmatch */
-	if (!CG_IsATeamGametype(cgs.gametype) || cgs.gametype == GT_TEAM ) {
+	if (CG_GametypeUsesFragLimit(cgs.gametype)) {
 		value = atoi( Info_ValueForKey( info, "fraglimit" ) );
 		if ( value ) {
 			UI_DrawProportionalString( 320, y, va( "fraglimit %i", value ),
@@ -301,7 +301,7 @@ void CG_DrawInformation( void ) {
 			y += PROP_HEIGHT;
 		}
 	}
-	else {
+	else /* (CG_GametypeUsesCaptureLimit(cgs.gametype)) */ {
 		value = atoi( Info_ValueForKey( info, "capturelimit" ) );
 		if ( value ) {
 			UI_DrawProportionalString( 320, y, va( "capturelimit %i", value ),