+++ code/cgame/cg_local.h
@@ -2046,4 +2046,6 @@ qboolean CG_IsATeamGametype(int check);	/* Whether the gametype is team-based or
 qboolean CG_UsesTeamFlags(int check);	/* Whether the gametype uses the red and blue flags. */
 qboolean CG_UsesTheWhiteFlag(int check);	/* Whether the gametype uses the neutral flag. */
 qboolean CG_IsARoundBasedGametype(int check);	/* Whether the gametype uses the neutral flag. */
+qboolean CG_GametypeUsesFragLimit(int check);	/* Whether the gametype uses a frag-based scoring system. */
+qboolean CG_GametypeUsesCaptureLimit(int check);	/* Whether the gametype uses a capture-based scoring system. */
 /* /Neon_Knight */